chore: Rename to types
@BasixKOR:
관점은 동의합니다만, 그 대안으로 domain이 타입을 저장하는 곳의 이름으로서 맞을 지는 잘 모르겠네요... types 같은 이름은 어떨까요?
@x86chi:
알겠습니다 😁
