+++ src/control.ts
@@ -1,7 +1,7 @@
 import { sendID, loadData } from './api'
-import { TypeLoad } from './domain/state'
+import { TypeLoad } from './types/state'
 import { fetchData } from './api'
-import { UserNoPw } from './domain/models'
+import { UserNoPw } from './types/models'
 
 export const submitID = async (mailid: string) => {
   try {
